backend/requirements.txt
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
gunicorn==23.0.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.10.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-sat==1.8.dev6
pytz==2024.2
PyYAML==5.4.1
recordclass==0.21
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.6
systemd-python==234
threadpoolctl==3.5.0
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
